# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Pemuda Surau Al-Abqori API
  version: 1.0.0
  description: |
    REST API for the Surau Al-Abqori youth programme: prayer attendance
    (face-verified + geofenced), Quran activity, merits, programs and the
    public dashboard.

    ## Authentication
    Most endpoints require a JWT obtained from `POST /api/auth/verify-otp`.
    Send it as `Authorization: Bearer <token>`.

    ## Authorisation
    Beyond "logged in", endpoints are gated by **capabilities** rather than raw
    roles (see `server/src/utils/roles.js`). Admins implicitly hold every
    capability. The `x-capability` extension on each operation records the
    requirement.

    ## Geofencing
    Check-in, activity recording and new registrations are only accepted within
    a radius of the surau. Coordinates are supplied by the client.
servers:
  - url: /api
    description: Same origin (proxied by Nginx / the Vite dev server)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Authentication required
        code:
          type: string
          description: Optional machine-readable code (e.g. REGISTRATION_TOO_FAR).
          example: REGISTRATION_TOO_FAR
        retryAfterMinutes:
          type: integer
          description: Present on 429 responses (e.g. the Quran log cooldown).
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        phone: { type: string, nullable: true, example: '+60123456789' }
        fullName: { type: string, example: Ahmad bin Zainal }
        birthDate: { type: string, format: date, nullable: true }
        age: { type: integer, nullable: true, example: 17 }
        gender: { type: string, nullable: true, enum: [male, female] }
        address: { type: string, nullable: true }
        role: { type: string, enum: [youth, admin, parent, teacher, ajk] }
        capabilities:
          type: array
          items: { type: string }
          example: [checkIn, viewMembers]
        hasFace: { type: boolean }
        telegramLinked: { type: boolean }
        guardianId: { type: string, format: uuid, nullable: true }
        isDependent: { type: boolean }
        canSelfCheckIn: { type: boolean }
        createdAt: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Rate limited.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /auth/request-otp:
    post:
      tags: [Auth]
      summary: Send a login code to a phone
      description: |
        Public. Rate limited (10 requests / 10 minutes). Delivery prefers the
        user's linked Telegram chat, then SMS, then the server console.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: '0123456789' }
      responses:
        '200':
          description: Code dispatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  channel: { type: string, enum: [telegram, sms, console] }
                  needsAdminHelp: { type: boolean }
                  telegramAvailable: { type: boolean }
                  telegramLinked: { type: boolean }
                  botUrl: { type: string }
                  devCode:
                    type: string
                    description: Only returned when NODE_ENV is not production.
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify a code and receive a JWT
      description: |
        Public. Creates the account if the phone is new — in which case the
        request must come from within the registration radius of the surau
        (default 1.5 km), so `latitude`/`longitude` are required for new users.
        Existing members may log in from anywhere.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string, example: '123456' }
                latitude: { type: number, description: Required only for a NEW account. }
                longitude: { type: number, description: Required only for a NEW account. }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
                  isNewUser: { type: boolean }
                  needsProfile: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403':
          description: |
            Deactivated account, a phone-less dependent account, or a new
            registration outside the allowed area (`code` is
            `REGISTRATION_TOO_FAR` / `REGISTRATION_LOCATION_REQUIRED`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ----------------------------------------------------------------- Users
...